requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python==4.9.0.80
pystray==0.19.5
Pillow==10.2.0
python-dotenv==1.0.1
easyocr==1.7.2
cv2-enumerate-cameras==1.3.3
requests==2.31.0
numpy==1.26.4
